Prevent tasks from being loaded twice
We want to set the `:scm` to `:rsync` to prevent loading the default SCM deploy
tasks.

Capistrano will auto-load the SCM in 'capistrano/setup'. By setting the `:scm`
to `:rsync` and loading 'capistrano/rsync' in the Capfile, the tasks would be
loaded twice.

By moving the tasks out into its separate file (strategy/rsync.rb) and using
`require` to load it in 'capistrano/rsync' we avoid having the tasks being
loaded twice.

Setting the `:scm` to `:rsync` is obviously a ugly hack, since Rsync is not a
SCM, but it is the only work around I can see for now, until
capistrano/capistrano#929 is resolved. It essentially works by setting the SCM
to be a no-op.